chore: make codebase clippy-clean
Fix clippy warnings across all crates to improve code quality:
- Add Default impls for types with new() methods (GpsPoint, KalmanState, DrState, etc.)
- Replace manual implementations with std methods (div_ceil, is_multiple_of, clamp)
- Remove unnecessary casts and improve type safety
- Fix collapsible if statements and redundant patterns
- Use vec![] macro instead of Vec::new() + push
- Replace filter_map(Result::ok) with map_while(Result::ok)
- Fix test loops to use enumerate() properly
- Add #[allow()] attributes for intentional patterns

All lib targets now pass clippy with -D warnings.
Remaining warnings are only in test files (minor style issues).

+++ crates/pico2-firmware/src/recovery_trigger.rs
@@ -18,7 +18,7 @@ use shared::DistCm;
 pub fn should_trigger_recovery(s_cm: DistCm, prev_s_cm: DistCm) -> bool {
     // GPS jump > 200 m (per spec Section 15.1)
     // Triggers recovery when position jumps > 200 m between consecutive fixes
-    let jump_distance = s_cm.abs_diff(prev_s_cm) as u32;
+    let jump_distance = s_cm.abs_diff(prev_s_cm);
     jump_distance > 20000
 }
