+++ crates/pipeline/detection/src/state_machine.rs
@@ -72,7 +72,7 @@ impl StopState {
         let d_to_stop = (s_cm - stop_progress).abs();
 
         // Track previous distance for re-acquisition detection
-        let current_distance = (stop_progress as i32) - (s_cm as i32);
+        let current_distance = stop_progress - s_cm;
         self.previous_distance_cm = Some(current_distance);
 
         match self.fsm_state {
